requirements.txt
google-cloud-speech==2.16.2
google-cloud-language==2.6.1
google-cloud-bigquery==3.4.2
functions-framework==3.2.0